// frontend/src/app/core/services/api/model/competitionDisciplineSchemeRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request to assign a discipline to a competition
 */
export interface CompetitionDisciplineSchemeRequest { 
    competitionId: string;
    disciplineId: string;
}